[noup] zephyr: nan: Cast BIT() results in nan_util.c to fix format warnings
Zephyr's own BIT() macro (zephyr/sys/util.h, "(1UL << (n))") wins over
hostap's own definition (src/utils/common.h, "(1U << (x))") by
include order in the Zephyr build, so these "%u"-based
wpa_scnprintf() calls end up passing an unsigned long where an
unsigned int is expected on 32-bit native_sim builds. This does not
reproduce on vanilla upstream, where hostap's own BIT() is always in
effect and already matches "%u".

Cast the values to unsigned int explicitly.

+++ src/nan/nan_util.c
@@ -1639,9 +1639,9 @@ int nan_peer_dump_sched_to_buf(struct nan_peer_schedule *sched,
 
 			ret = wpa_scnprintf(pos, end - pos,
 					    "\t\tbitmap: period=%u duration=%u offset=%u ",
-					    BIT(6 + chan->tbm.period),
-					    BIT(4 + chan->tbm.duration),
-					    16 * chan->tbm.offset);
+					    (unsigned int) BIT(6 + chan->tbm.period),
+					    (unsigned int) BIT(4 + chan->tbm.duration),
+					    (unsigned int) (16 * chan->tbm.offset));
 			if (os_snprintf_error(end - pos, ret))
 				goto err;
 			pos += ret;
@@ -1665,9 +1665,9 @@ int nan_peer_dump_sched_to_buf(struct nan_peer_schedule *sched,
 
 		ret = wpa_scnprintf(pos, end - pos,
 				    "\tndc: period=%u duration=%u offset=%u bitmap=",
-				    BIT(6 + map->ndc.period),
-				    BIT(4 + map->ndc.duration),
-				    16 * map->ndc.offset);
+				    (unsigned int) BIT(6 + map->ndc.period),
+				    (unsigned int) BIT(4 + map->ndc.duration),
+				    (unsigned int) (16 * map->ndc.offset));
 		if (os_snprintf_error(end - pos, ret))
 			goto err;
 		pos += ret;
